django-venv/requirements.txt
astroid==2.2.5
beautifulsoup4==4.7.1
colorama==0.4.1
Django==2.2
django-filter==2.1.0
djangorestframework==3.9.2
isort==4.3.17
lazy-object-proxy==1.3.1
lxml==4.3.3
Markdown==3.1
mccabe==0.6.1
pylint==2.3.1
pytz==2019.1
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
typed-ast==1.3.1
wrapt==1.11.1
